Rename and refactor environments handler in sap systems handlers, update tests
